Unclosed array expression should throw exception
In the current implementation we do not take into account that people may try
the following notation:

    `@var array(string => string|null)`

At the moment this is interpreted as an Array Expression but fails because the
space causes the parser to see this as `array(string`, meaning an unclosed
array expression.

This change introduces a new RuntimeException that is thrown in this situation,
it is up to the consumer of this library to deal with this situation however it
deems best.


Former-commit-id: 9c0562f9af69b06b002b017a3f0672bbfaba349f